handled offline and network issues + network timeout

+++ app/src/main/java/com/mirhoseini/marketprice/ui/main/model/MainInteractorImpl.java
@@ -29,31 +29,45 @@ public class MainInteractorImpl implements MainInteractor {
 
     @Override
     public void loadMarketPrice(TimeSpan timeSpan, final OnMainNetworkFinishedListener listener) {
-        mNetworkHelper.loadMarketPriceValues(timeSpan, new OnNetworkFinishedListener<RestMarketPrice>() {
-            @Override
-            public void onSuccess(TimeSpan timeSpan, RestMarketPrice restResponse) {
-                if (listener != null)
-                    listener.onNetworkSuccess(timeSpan, restResponse);
-            }
+        if (mNetworkHelper != null) {//check if object is not destroyed
 
-            @Override
-            public void onError(TimeSpan timeSpan, Throwable throwable) {
-                if (listener != null)
-                    listener.onNetworkError(timeSpan, throwable);
-            }
-        });
+            mNetworkHelper.loadMarketPriceValues(timeSpan, new OnNetworkFinishedListener<RestMarketPrice>() {
+                @Override
+                public void onSuccess(TimeSpan timeSpan, RestMarketPrice restResponse) {
+                    if (listener != null)
+                        listener.onNetworkSuccess(timeSpan, restResponse);
+                }
+
+                @Override
+                public void onError(TimeSpan timeSpan, Throwable throwable) {
+                    if (listener != null)
+                        listener.onNetworkError(timeSpan, throwable);
+                }
+            });
+
+        }
     }
 
     @Override
     public List<PriceValue> loadPriceValuesFromDatabase(TimeSpan timeSpan) {
-        List<PriceValue> items = mDatabaseHelper.loadPriceValues(timeSpan.getPosition());
+        List<PriceValue> items = null;
+
+        if (mDatabaseHelper != null) {//check if object is not destroyed
+
+            items = mDatabaseHelper.loadPriceValues(timeSpan.getPosition());
+
+        }
 
         return items;
     }
 
     @Override
     public void deletePriceValues(TimeSpan timeSpan) {
-        mDatabaseHelper.deletePriceValues(timeSpan.getPosition());
+        if (mDatabaseHelper != null) {//check if object is not destroyed
+
+            mDatabaseHelper.deletePriceValues(timeSpan.getPosition());
+
+        }
     }
 
     @Override
